fix: Rename lib -> libs
