+++ src/libs/mail/ses.test.ts
@@ -5,7 +5,7 @@ import {
 } from '@aws-sdk/client-sesv2'
 import { mockClient } from 'aws-sdk-client-mock'
 
-import { SESApiError } from '@/lib/errors'
+import { SESApiError } from '@/libs/errors'
 
 import { type Mail } from '@/types/Mail'
